+++ src/app/core/@ngrx/app.state.ts
@@ -1,0 +1,7 @@
+import { ProductsState } from './products';
+import { RouterStateUrl } from './router';
+
+export interface AppState {
+  products: ProductsState;
+  router: RouterStateUrl;
+}