+++ src/app/core/@ngrx/index.ts
@@ -1,0 +1,3 @@
+export {AppState} from './app.state';
+export * from './products';
+export * from './router';