+++ src/app/core/@ngrx/products/index.ts
@@ -1,0 +1,5 @@
+import * as ProductsActions from './products.actions';
+export {ProductsState} from './products.state';
+export * from './products.effects';
+export {ProductsActions};
+export * from './products.selectors';