+++ src/app/core/@ngrx/router/index.ts
@@ -1,0 +1,7 @@
+import * as RouterActions from './router.actions';
+export {RouterActions};
+export * from './router.custom-serializer';
+export * from './router.reducer';
+export * from './router.state';
+export * from './router.selectors';
+export * from './router.effects';